+++ packages/components-vue/src/router/index.ts
@@ -81,29 +81,29 @@ export const routes: RouteType[] = [
     component: fromExamples.FlyoutExample,
   },
   {
-    path: '/flyout-navigation-example',
-    name: 'Flyout Navigation Example',
-    component: fromExamples.FlyoutNavigationExample,
+    path: '/flyout-multilevel-example',
+    name: 'Flyout Multilevel Example',
+    component: fromExamples.FlyoutMultilevelExample,
   },
   {
-    path: '/flyout-navigation-example-active-identifier',
-    name: 'Flyout Navigation Example Active Identifier',
-    component: fromExamples.FlyoutNavigationExampleActiveIdentifier,
+    path: '/flyout-multilevel-example-active-identifier',
+    name: 'Flyout Multilevel Example Active Identifier',
+    component: fromExamples.FlyoutMultilevelExampleActiveIdentifier,
   },
   {
-    path: '/flyout-navigation-example-custom-content',
-    name: 'Flyout Navigation Example Custom Content',
-    component: fromExamples.FlyoutNavigationExampleCustomContent,
+    path: '/flyout-multilevel-example-custom-content',
+    name: 'Flyout Multilevel Example Custom Content',
+    component: fromExamples.FlyoutMultilevelExampleCustomContent,
   },
   {
-    path: '/flyout-navigation-basic',
-    name: 'FlyoutNavigationBasic',
-    component: fromPages.FlyoutNavigationBasicPage,
+    path: '/flyout-multilevel-basic',
+    name: 'FlyoutMultilevelBasic',
+    component: fromPages.FlyoutMultilevelBasicPage,
   },
   {
-    path: '/flyout-navigation-prefixed',
-    name: 'FlyoutNavigationPrefixed',
-    component: fromPages.FlyoutNavigationPrefixedPage,
+    path: '/flyout-multilevel-prefixed',
+    name: 'FlyoutMultilevelPrefixed',
+    component: fromPages.FlyoutMultilevelPrefixedPage,
   },
   {
     path: '/inline-notification-example-action-button',
